refactor(core): 重构应用初始化逻辑
- 引入 AppBuilder 模式替代全局静态变量注册插件与适配器
- 修改 init_app 函数签名以接收插件和适配器列表作为参数
- 调整系统信息结构体字段类型提高精度
- 升级 HTTP 相关依赖版本并引入 reqwest 支持
- 移除旧的 ADAPTER_DIR 和相关路径常量引用
- 优化路径管理逻辑支持设置工作目录
- 修复测试用例中断言逻辑及数据类型匹配问题
- 清理无用导入和冗余代码提升可读性
- 统一错误处理流程增强程序健壮性
- 调整 Cargo.lock 文件中多个包的版本与校验值
- 更新 puniyu 主函数使用新的构建器模式启动应用

+++ crates/puniyu_common/src/system.rs
@@ -77,11 +77,11 @@ pub struct BotStatusInfo {
 	/// 进程PID
 	pub pid: Pid,
 	/// CPU使用率
-	pub cpu_usage: Option<u8>,
+	pub cpu_usage: Option<f32>,
 	/// 内存使用率
-	pub memory_usage: Option<u8>,
+	pub memory_usage: Option<f32>,
 	/// 已用内存(单位: MB)
-	pub used_memory: f32,
+	pub used_memory: f64,
 	/// 运行时间
 	pub run_time: u64,
 }
